Change to the cpacs definition of the kinks

+++ src/CCPACSCurvePointListXYZ.h
@@ -19,16 +19,16 @@ public:
 
     /// Returns the parameters of the kinks. If no kinks are defined
     /// this list is empty.
-    TIGL_EXPORT const std::vector<double>& GetKinksAsVector() const;
+    TIGL_EXPORT const std::vector<unsigned int>& GetKinksAsVector() const;
 
     TIGL_EXPORT const std::vector<CTiglPoint>& AsVector() const;
     TIGL_EXPORT void SetValue(int index, const CTiglPoint& point);
     TIGL_EXPORT void SetAsVector(const std::vector<CTiglPoint>& points);
 
 private:
-    void BuildKinks(std::vector<double>& kinks) const;
+    void BuildKinks(std::vector<unsigned int>& kinks) const;
 
-    Cache<std::vector<double>, CCPACSCurvePointListXYZ> m_kinksVec;
+    Cache<std::vector<unsigned int>, CCPACSCurvePointListXYZ> m_kinksVec;
     // cache
     std::vector<CTiglPoint> m_vec;
 };
